web服务端对设备的控制领域服务基本完成，还需要进一步抽象改进

+++ src/Equipment/BXJG.Equipment.Share/Protocol/OxygenChamberPackage.cs
@@ -50,7 +50,7 @@ namespace BXJG.Equipment.Protocol
     public class cykz : OxygenChamberPackage
     {
         public bool Add { get; set; }
-        public byte Value { get; set; }
+        public short Value { get; set; }
     }
     /// <summary>
     /// 状态上报
